+++ src/main/java/com/notepad/NotepadApp/repository/NoteRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NoteRepository extends JpaRepository<Note, Long> {
+public interface NoteRepository extends JpaRepository<Note, Integer> {
 
 //    Note findByDate(long id);
 }
